simplify, a lot
