+++ cmd/cmd.go
@@ -3,15 +3,31 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"io"
 	"os"
 	"os/signal"
+	"strings"
 	"syscall"
 
+	"github.com/knadh/koanf"
+	"github.com/knadh/koanf/parsers/toml"
+	"github.com/knadh/koanf/providers/file"
 	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
 	"golang.org/x/sync/errgroup"
+)
 
-	"github.com/umee-network/liquidator/core"
+const (
+	envVariablePass = "LIQUIDATOR_PASSWORD"
+	flagConfigPath  = "config-file"
+	logLevelJSON    = "json"
+	logLevelText    = "text"
+)
+
+var (
+	konfig   *koanf.Koanf
+	logger   *zerolog.Logger
+	password string
 )
 
 var rootCmd = &cobra.Command{
@@ -32,26 +48,23 @@ func Execute() {
 }
 
 func liquidatorCmdHandler(cmd *cobra.Command, _ []string) error {
-	lock.Lock()
-	launched = true // prevents further calls to Init
-	lock.Unlock()
 
 	configPath, err := cmd.Flags().GetString(flagConfigPath)
 	if err != nil {
 		return err
 	}
 
-	konfig, err := getConfig(configPath)
+	konfig, err = loadConfig(configPath)
 	if err != nil {
 		return err
 	}
 
-	logger, err := getLogger(konfig)
+	logger, err = getLogger(konfig)
 	if err != nil {
 		return err
 	}
 
-	password, err := getPass()
+	password, err = getPassword()
 	if err != nil {
 		return err
 	}
@@ -64,12 +77,65 @@ func liquidatorCmdHandler(cmd *cobra.Command, _ []string) error {
 
 	g.Go(func() error {
 		// returns on context cancelled
-		return core.Start(ctx, logger, konfig, password, cancel)
+		return startLiquidator(ctx, cancel)
 	})
 
 	// Block main process until all spawned goroutines have gracefully exited and
 	// signal has been captured in the main process or if an error occurs.
 	return g.Wait()
+}
+
+// getPassword reads the keyring password from an environment variable
+func getPassword() (string, error) {
+	pass := os.Getenv(envVariablePass)
+	if pass == "" {
+		return "", fmt.Errorf("empty keyring password")
+	}
+	return pass, nil
+}
+
+// loadConfig returns a koanf configuration loaded from a specified filepath
+func loadConfig(path string) (*koanf.Koanf, error) {
+	var k = koanf.New(".")
+
+	// Load toml config from specified file path
+	f := file.Provider(path)
+	if err := k.Load(f, toml.Parser()); err != nil {
+		return nil, err
+	}
+
+	return k, nil
+}
+
+// getLogger returns a zerolog logger configured from the "log.level" and "log.format"
+// fields of a koanf config. Log format should be "json" or "text".
+func getLogger(konfig *koanf.Koanf) (*zerolog.Logger, error) {
+
+	logLvl, err := zerolog.ParseLevel(konfig.String("log.level"))
+	if err != nil {
+		return nil, err
+	}
+
+	logFormat := strings.ToLower(konfig.String("log.format"))
+
+	var logWriter io.Writer
+	if strings.ToLower(logFormat) == logLevelText {
+		logWriter = zerolog.ConsoleWriter{Out: os.Stderr}
+	} else {
+		logWriter = os.Stderr
+	}
+
+	switch logFormat {
+	case logLevelJSON:
+		logWriter = os.Stderr
+	case logLevelText:
+		logWriter = zerolog.ConsoleWriter{Out: os.Stderr}
+	default:
+		return nil, fmt.Errorf("invalid logging format: %s", logFormat)
+	}
+
+	l := zerolog.New(logWriter).Level(logLvl).With().Timestamp().Logger()
+	return &l, nil
 }
 
 // trapSignal will listen for any OS signal and invoke Done on the main
